% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{account_profile}
\alias{account_profile}
\title{Account Profile}
\usage{
account_profile()
}
\description{
Returns information about the Shodan account linked to this API key.
}
\examples{
account_profile()
}
\references{
\url{https://developer.shodan.io/api}
}

