% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{api_info}
\alias{api_info}
\title{API Plan Information}
\usage{
api_info(ips = NULL)
}
\description{
Returns information about the API plan belonging to the given API key.
}
\examples{
my_ip()
}
\references{
\url{https://developer.shodan.io/api}
}
