% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/host_count.r
\name{host_count}
\alias{host_count}
\title{Search Shodan without Results}
\usage{
host_count(query = NULL, facets = NULL)
}
\arguments{
\item{query}{Shodan search query. The provided string is used to search the
database of banners in Shodan, with the additional option to provide
filters inside the search query using a \code{"filter:value"} format. For
example, the following search query would find Apache webservers
located in Germany: \code{"apache country:DE"}. See
\url{https://developer.shodan.io/api} for all supported search filters.}

\item{facets}{A comma-separated list of properties to get summary information on.
Property names can also be in the format of \code{"property:count"},
where \code{"count"} is the number of facets that will be returned for
a property (i.e. \code{"country:100"} to get the top 100 countries for
a search query). See \url{https://developer.shodan.io/api} for all
supported facets.}
}
\description{
This method behaves identical to "/shodan/host/search" with the only
difference that this method does not return any host results, it only
returns the total number of results that matched the query and any facet
information that was requested. As a result this method does not consume
query credits.
}
\details{
Returns all services that have been found on the given host IP.
}
\examples{
host_count("41.21.249.170")
}
\references{
\url{https://developer.shodan.io/api}
}

