% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/host_info.r
\name{host_info}
\alias{host_info}
\title{Host Information}
\usage{
host_info(ip, history = FALSE, minify = FALSE)
}
\arguments{
\item{ip}{Host IP address}

\item{history}{\code{TRUE} == all historical banners should be returned
(default: \code{FALSE})}

\item{minify}{\code{TRUE} == only return the list of ports and the general
host information, no banners. (default: \code{FALSE})}
}
\description{
Returns all services that have been found on the given host IP.
}
\examples{
host_info("41.21.249.170")
}
\references{
\url{https://developer.shodan.io/api}
}

