% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{http_headers}
\alias{http_headers}
\title{Shows the HTTP headers that your client sends when connecting to a webserver.}
\usage{
http_headers(ips = NULL)
}
\description{
Shows the HTTP headers that your client sends when connecting to a webserver.
}
\examples{
http_headers()
}
\references{
\url{https://developer.shodan.io/api}
}
