% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{my_ip}
\alias{my_ip}
\title{My IP Address}
\usage{
my_ip(ips = NULL)
}
\description{
Get your current IP address as seen from the Internet.
}
\examples{
my_ip()
}
\references{
\url{https://developer.shodan.io/api}
}
