% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{query_tags}
\alias{query_tags}
\title{List the most popular tags}
\usage{
query_tags(size = 10)
}
\arguments{
\item{size}{The number of tags to return (default: \code{10}).}
}
\description{
Use this method to obtain a list of popular tags for the saved search
queries in Shodan.
}
\examples{
query_tags()
}
\references{
\url{https://developer.shodan.io/api}
}

