% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dns.r
\name{resolve}
\alias{resolve}
\title{DNS Lookup}
\usage{
resolve(hostnames = NULL)
}
\arguments{
\item{hostnames}{Comma-separated list of hostnames; example
\code{"google.com,bing.com"}. If \code{length(hostnames) > 1}
the entire vector will be converted to a comma-separated parameter
to pass to Shodan.}
}
\description{
Look up the IP address for the provided list of hostnames.
}
\examples{
resolve("google.com,bing.com")
resolve(c("google.com", "bing.com"))
}
\references{
\url{https://developer.shodan.io/api}
}
