% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dns.r
\name{reverse}
\alias{reverse}
\title{Reverse DNS Lookup}
\usage{
reverse(ips = NULL)
}
\arguments{
\item{ips}{Comma-separated list of IP addresses; example
\code{"74.125.227.230,204.79.197.200"}. If \code{length(ips) > 1}
the entire vector will be converted to a comma-separated parameter
to pass to Shodan.}
}
\description{
Look up the hostnames that have been defined for the given list of IP
addresses.
}
\examples{
reverse("74.125.227.230,204.79.197.200")
reverse(c("74.125.227.230", "204.79.197.200"))
}
\references{
\url{https://developer.shodan.io/api}
}

