% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shodan-package.r
\docType{package}
\name{shodan}
\alias{shodan}
\alias{shodan-package}
\title{Access the Shodan API from R}
\description{
See \url{https://developer.shodan.io/api/requirements} for information
on the Shodan API
}
\author{
Bob Rudis (@hrbrmstr)
}

