% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shodan.r
\name{shodan_api_key}
\alias{shodan_api_key}
\title{Get or set SHODAN_API_KEY value}
\usage{
shodan_api_key(force = FALSE)
}
\arguments{
\item{force}{Force setting a new Shodan API key for the current environment?}
}
\value{
atomic character vector containing the Shodan API key
}
\description{
The API wrapper functions in this package all rely on a Shodan API
key residing in the environment variable \code{SHODAN_API_KEY}. The
easiest way to accomplish this is to set it in the `.Renviron` file in your
home directory.
}
