% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploits.r
\name{shodan_exploit_search}
\alias{shodan_exploit_search}
\title{Search for Exploits}
\usage{
shodan_exploit_search(query = NULL, facets = NULL, page = 1)
}
\arguments{
\item{query}{Search query used to search the database of known exploits. See
\url{https://developer.shodan.io/api/exploits/rest} for all supported
search filters.}

\item{facets}{A comma-separated list of properties to get summary information on.
The following facets are currently supported: "\code{author}",
"\code{platform}", "\code{port}", "\code{source}" and "\code{type}.
If \code{length(facets) > 1)} this function will
concatenate the vector with commas to send to Shodan.}

\item{page}{The page number to page through results \code{100} at a time
(default: \code{1})}
}
\description{
Search across a variety of data sources for exploits and use facets to
get summary information.
}
\references{
\url{https://developer.shodan.io/api/exploits/rest}
}

