% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ports.r
\name{shodan_ports}
\alias{shodan_ports}
\title{List all ports that Shodan is crawling on the Internet.}
\usage{
shodan_ports()
}
\description{
This method returns a list of port numbers that the crawlers are looking for.
}
\examples{
shodan_ports()
}
\references{
\url{https://developer.shodan.io/api}
}

