% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ports.r
\name{shodan_protocols}
\alias{shodan_protocols}
\title{List all protocols that can be used when performing on-demand Internet
scans via Shodan.}
\usage{
shodan_protocols()
}
\description{
This method returns an object containing all the protocols that can be
used when launching an Internet scan.
}
\examples{
shodan_protocols()
}
\references{
\url{https://developer.shodan.io/api}
}
