% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.r
\name{shodan_query_list}
\alias{shodan_query_list}
\title{List the saved search queries}
\usage{
shodan_query_list(page = 1, sort = NULL, order = NULL)
}
\arguments{
\item{page}{Page number to iterate over results; each page contains 10 items}

\item{sort}{Sort the list based on a property. Possible values are:
"\code{votes}", "\code{timestamp}"}

\item{order}{Whether to sort the list in ascending or descending order.
Possible values are: "\code{asc}", "\code{desc}"}
}
\description{
Use this method to obtain a list of search queries that users have saved
in Shodan.
}
\examples{
shodan_query_list()
}
\references{
\url{https://developer.shodan.io/api}
}

