% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.r
\name{shodan_query_search}
\alias{shodan_query_search}
\title{Search the directory of saved search queries.}
\usage{
shodan_query_search(query = NULL, page = 1)
}
\arguments{
\item{query}{What to search for in the directory of saved search queries.}

\item{page}{Page number to iterate over results; each page contains 10 items}
}
\description{
Use this method to search the directory of search queries that users have
saved in Shodan.
}
\examples{
shodan_query_search()
}
\references{
\url{https://developer.shodan.io/api}
}
