% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scan.r
\name{shodan_scan}
\alias{shodan_scan}
\title{Request Shodan to crawl an IP/ netblock}
\usage{
shodan_scan(ips = NULL)
}
\arguments{
\item{ips}{A comma-separated list of IPs or netblocks (in CIDR notation)
that should get crawled. If \code{length(ips) > 1} this funtion
will concatenate the vector into a comma-separated string to
send to Shodan.}
}
\description{
Use this method to request Shodan to crawl a network.
}
\details{
This method uses API scan credits: 1 IP consumes 1 scan credit.
You must have a paid API plan (either one-time payment or subscription)
in order to use this method.
}
\examples{
shodan_scan("216.0.0.0/16")
}
\references{
\url{https://developer.shodan.io/api}
}

