% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scan.r
\name{shodan_scan_internet}
\alias{shodan_scan_internet}
\title{Crawl the Internet for a specific port and protocol using Shodan}
\usage{
shodan_scan_internet(port = NULL, protocol = NULL)
}
\arguments{
\item{port}{The port that Shodan should crawl the Internet for.}

\item{protocol}{The name of the protocol that should be used to interrogate
the port. Run \link{shodan_protocols} for a list of supported protocols.}
}
\description{
Use this method to request Shodan to crawl the Internet for a specific port.
}
\details{
This method is restricted to security researchers and companies with a
Shodan Data license. To apply for access to this method as a researcher,
please email <jmath@shodan.io> with information about your project. Access
is restricted to prevent abuse.
}
\references{
\url{https://developer.shodan.io/api}
}

