% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.r
\name{shodan_search}
\alias{shodan_search}
\title{Search Shodan}
\usage{
shodan_search(query = NULL, facets = NULL, page = 1, minify = TRUE)
}
\arguments{
\item{query}{Shodan search query. The provided string is used to search the
database of banners in Shodan, with the additional option to provide
filters inside the search query using a \code{"filter:value"} format. For
example, the following search query would find Apache webservers
located in Germany: \code{"apache country:DE"}. See
\url{https://developer.shodan.io/api} for all supported search filters.}

\item{facets}{A comma-separated list of properties to get summary information on.
Property names can also be in the format of \code{"property:count"},
where \code{"count"} is the number of facets that will be returned for
a property (i.e. \code{"country:100"} to get the top 100 countries for
a search query). See \url{https://developer.shodan.io/api} for all
supported facets. If \code{length(facets) > 1)} this function will
concatenate the vector with commas to send to Shodan.}

\item{page}{The page number to page through results \code{100} at a time
(default: \code{1})}

\item{minify}{\code{TRUE} == truncate some of the larger fields (default: \code{TRUE})}
}
\description{
Search Shodan using the same query syntax as the website and use facets to
get summary information for different properties.
}
\details{
This method may use API query credits depending on usage. If any of the
following criteria are met, your account will be deducated 1 query credit:

\itemize{
  \item{The search query contains a filter.}
  \item{Accessing results past the 1st page using the "page". For every 100
        results past the 1st page 1 query credit is deducted.}
}
}
\examples{
shodan_search("apache geo:42.9693,-74.1224")
}
\references{
\url{https://developer.shodan.io/api}
}

