% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.r
\name{shodan_search_tokens}
\alias{shodan_search_tokens}
\title{Break the search query into tokens}
\usage{
shodan_search_tokens(query = NULL)
}
\arguments{
\item{query}{Shodan search query. The provided string is used to search the
database of banners in Shodan, with the additional option to provide
filters inside the search query using a \code{"filter:value"} format. For
example, the following search query would find Apache webservers
located in Germany: \code{"apache country:DE"}. See
\url{https://developer.shodan.io/api} for all supported search filters.}
}
\description{
This method lets you determine which filters are being used by the query
string and what parameters were provided to the filters.
}
\examples{
shodan_search_tokens("iis net:216.0.0.0/16")
}
\references{
\url{https://developer.shodan.io/api}
}
