% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ports.r
\name{shodan_services}
\alias{shodan_services}
\title{List all services that Shodan crawls}
\usage{
shodan_services()
}
\description{
This method returns an object containing all the services that the Shodan
crawlers look at. It can also be used as a quick and practical way to
resolve a port number to the name of a service.
}
\examples{
shodan_services()
}
\references{
\url{https://developer.shodan.io/api}
}
